Harden and refactor object grouping
